moving stuff around + dll compile fix
