+++ thorin/CMakeLists.txt
@@ -104,6 +104,6 @@ target_include_directories(libthorin PUBLIC
 )
 target_link_libraries(libthorin PUBLIC absl::flat_hash_map absl::flat_hash_set
                                        absl::node_hash_map absl::node_hash_set)
-#if(MSVC)
-    #target_compile_definitions(libthorin PUBLIC ABSL_CONSUME_DLL)
-#endif()
+if(MSVC AND BUILD_SHARED_LIBS)
+    target_compile_definitions(libthorin PUBLIC ABSL_CONSUME_DLL)
+endif()
