+++ thorin/def.h
@@ -10,6 +10,7 @@
 #include "thorin/util/array.h"
 #include "thorin/util/cast.h"
 #include "thorin/util/hash.h"
+#include "thorin/util/container.h"
 #include "thorin/util/stream.h"
 
 namespace thorin {
