+++ thorin/util/hash.h
@@ -1,11 +1,6 @@
 #ifndef THORIN_UTIL_HASH_H
 #define THORIN_UTIL_HASH_H
 
-#include <absl/container/flat_hash_map.h>
-#include <absl/container/flat_hash_set.h>
-#include <absl/container/node_hash_map.h>
-#include <absl/container/node_hash_set.h>
-
 #include "thorin/util/types.h"
 
 namespace thorin {
@@ -118,29 +113,6 @@ inline hash_t hash_begin() { return FNV1::offset; }
 ///@{
 hash_t hash(const char*);
 hash_t hash(std::string_view);
-///@}
-
-/// @name maps/sets based upon gid
-///@{
-template<class T>
-struct GIDHash {
-    size_t operator()(T p) const { return murmur3(p->gid()); };
-};
-
-template<class T>
-struct GIDEq {
-    bool operator()(T a, T b) const { return a->gid() == b->gid(); }
-};
-
-template<class T>
-struct GIDLt {
-    bool operator()(T a, T b) const { return a->gid() < b->gid(); }
-};
-
-template<class K, class V> using GIDMap     = absl::flat_hash_map<K, V, GIDHash<K>, GIDEq<K>>;
-template<class K>          using GIDSet     = absl::flat_hash_set<K,    GIDHash<K>, GIDEq<K>>;
-template<class K, class V> using GIDNodeMap = absl::node_hash_map<K, V, GIDHash<K>, GIDEq<K>>;
-template<class K>          using GIDNodeSet = absl::node_hash_set<K,    GIDHash<K>, GIDEq<K>>;
 ///@}
 
 } // namespace thorin
